pyproject.toml
[project]
name = "strands-agents-demo"
version = "0.1.0"
description = "Strands Agents demo with Azure OpenAI integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0"
]
